Implemented Obstacle Foundation
Obstacles should be ready to use as Terrain supports them and the logic
is there. Just needs collision handling, so no obstacles get spawned
currently.

Also tweaked Terrain so there's no more bumps. Need to implement 'Set
Pieces' for terrain for the pot holes and such that we talked about if
we pursue that.